fsi-slide-studio/requirements.txt
streamlit>=1.42.0
claude-agent-sdk>=0.1.35
python-dotenv>=1.0.0
pyyaml>=6.0
